# mobile/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
      "ForecastHorizon": {
        "enum": [
          "30m",
          "1d",
          "5d",
          "7d",
          "1mo",
          "6mo",
          "1y"
        ],
        "title": "ForecastHorizon",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "StockDetailForecast": {
        "properties": {
          "candlesticks": {
            "items": {
              "$ref": "#/components/schemas/StockDetailForecastCandlestick"
            },
            "title": "Candlesticks",
            "type": "array"
          },
          "freshnessLabel": {
            "title": "Freshnesslabel",
            "type": "string"
          },
          "generatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generatedat"
          },
          "linePoints": {
            "items": {
              "$ref": "#/components/schemas/StockDetailForecastLinePoint"
            },
            "title": "Linepoints",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "generatedAt",
          "freshnessLabel"
        ],
        "title": "StockDetailForecast",
        "type": "object"
      },
      "StockDetailForecastCandlestick": {
        "properties": {
          "close": {
            "title": "Close",
            "type": "number"
          },
          "high": {
            "title": "High",
            "type": "number"
          },
          "low": {
            "title": "Low",
            "type": "number"
          },
          "open": {
            "title": "Open",
            "type": "number"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "sequence",
          "timestamp",
          "open",
          "high",
          "low",
          "close"
        ],
        "title": "StockDetailForecastCandlestick",
        "type": "object"
      },
      "StockDetailForecastLinePoint": {
        "properties": {
          "expectedValue": {
            "title": "Expectedvalue",
            "type": "number"
          },
          "lowerBound": {
            "title": "Lowerbound",
            "type": "number"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "upperBound": {
            "title": "Upperbound",
            "type": "number"
          }
        },
        "required": [
          "sequence",
          "timestamp",
          "expectedValue",
          "lowerBound",
          "upperBound"
        ],
        "title": "StockDetailForecastLinePoint",
        "type": "object"
      },
      "StockDetailHorizonMetadata": {
        "properties": {
          "calendarBasis": {
            "enum": [
              "regular_market_trading_time",
              "calendar_period"
            ],
            "title": "Calendarbasis",
            "type": "string"
          },
          "expectedForecastPointCount": {
            "title": "Expectedforecastpointcount",
            "type": "integer"
          },
          "externalFactorWeightScale": {
            "title": "Externalfactorweightscale",
            "type": "number"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "newsWindowDays": {
            "title": "Newswindowdays",
            "type": "integer"
          },
          "pricePointBasis": {
            "$ref": "#/components/schemas/StockDetailPricePointBasis"
          },
          "timeBasis": {
            "enum": [
              "regular_market",
              "calendar_period"
            ],
            "title": "Timebasis",
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/ForecastHorizon"
          }
        },
        "required": [
          "value",
          "label",
          "timeBasis",
          "pricePointBasis",
          "calendarBasis",
          "newsWindowDays",
          "externalFactorWeightScale",
          "expectedForecastPointCount"
        ],
        "title": "StockDetailHorizonMetadata",
        "type": "object"
      },
      "StockDetailKeyFactor": {
        "properties": {
          "factorType": {
            "title": "Factortype",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "polarity": {
            "enum": [
              "positive",
              "negative",
              "neutral"
            ],
            "title": "Polarity",
            "type": "string"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "sourceId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceid"
          },
          "sourceReferenceType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcereferencetype"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          }
        },
        "required": [
          "factorType",
          "sourceReferenceType",
          "sourceId",
          "label",
          "value",
          "rationale",
          "polarity",
          "weight"
        ],
        "title": "StockDetailKeyFactor",
        "type": "object"
      },
      "StockDetailMarket": {
        "properties": {
          "dailyChange": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dailychange"
          },
          "dailyChangePercent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dailychangepercent"
          },
          "freshnessLabel": {
            "title": "Freshnesslabel",
            "type": "string"
          },
          "latestPrice": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestprice"
          },
          "observedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observedat"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "latestPrice",
          "dailyChange",
          "dailyChangePercent",
          "observedAt",
          "freshnessLabel"
        ],
        "title": "StockDetailMarket",
        "type": "object"
      },
      "StockDetailPrediction": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "direction": {
            "anyOf": [
              {
                "enum": [
                  "bullish",
                  "bearish",
                  "neutral"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "expectedChangePercent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedchangepercent"
          },
          "freshnessLabel": {
            "title": "Freshnesslabel",
            "type": "string"
          },
          "generatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generatedat"
          },
          "keyFactors": {
            "items": {
              "$ref": "#/components/schemas/StockDetailKeyFactor"
            },
            "title": "Keyfactors",
            "type": "array"
          },
          "riskLevel": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium",
                  "high"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risklevel"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "direction",
          "confidence",
          "expectedChangePercent",
          "riskLevel",
          "generatedAt",
          "freshnessLabel"
        ],
        "title": "StockDetailPrediction",
        "type": "object"
      },
      "StockDetailPricePointBasis": {
        "enum": [
          "trading_session"
        ],
        "title": "StockDetailPricePointBasis",
        "type": "string"
      },
      "StockDetailResponse": {
        "properties": {
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "forecast": {
            "$ref": "#/components/schemas/StockDetailForecast"
          },
          "horizon": {
            "$ref": "#/components/schemas/ForecastHorizon"
          },
          "horizonMetadata": {
            "$ref": "#/components/schemas/StockDetailHorizonMetadata"
          },
          "market": {
            "$ref": "#/components/schemas/StockDetailMarket"
          },
          "prediction": {
            "$ref": "#/components/schemas/StockDetailPrediction"
          },
          "stock": {
            "$ref": "#/components/schemas/StockDetailStock"
          }
        },
        "required": [
          "stock",
          "horizon",
          "horizonMetadata",
          "market",
          "forecast",
          "prediction",
          "disclaimer"
        ],
        "title": "StockDetailResponse",
        "type": "object"
      },
      "StockDetailStock": {
        "properties": {
          "companyName": {
            "title": "Companyname",
            "type": "string"
          },
          "exchange": {
            "title": "Exchange",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "companyName",
          "exchange"
        ],
        "title": "StockDetailStock",
        "type": "object"
      },
      "StockSearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/StockSearchResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "StockSearchResponse",
        "type": "object"
      },
      "StockSearchResult": {
        "properties": {
          "companyName": {
            "title": "Companyname",
            "type": "string"
          },
          "exchange": {
            "title": "Exchange",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "companyName",
          "exchange"
        ],
        "title": "StockSearchResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Trendwise API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/stocks/search": {
      "get": {
        "operationId": "searchStocks",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Stocks",
        "tags": [
          "stocks"
        ]
      }
    },
    "/stocks/{ticker}/detail": {
      "get": {
        "operationId": "getStockDetail",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ForecastHorizon",
              "default": "1d"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Stock Detail",
        "tags": [
          "stocks"
        ]
      }
    }
  }
}